pyproject.toml
[project]
name = "qr-popup"
version = "0.1.0"
description = "A small tool to generate and display a QR code in a popup window"
readme = "README.md"
requires-python = ">=3.7"
authors = [{ name="Your Name", email="you@example.com" }]
dependencies = [
    "qrcode[pil]",
    "pillow"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
